requirements.txt
fastapi==0.95.2
pydantic==1.10.7
uvicorn==0.22.0
SQLAlchemy==2.0.13
requests==2.30.0
psycopg2==2.9.6
alembic==1.11.0
Jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
